% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/curvFilter-methods.R
\name{addName,curv1Filter,character-method}
\alias{addName,curv1Filter,character-method}
\alias{addName,curv1Filter,logical-method}
\alias{addName,curv2Filter,character-method}
\alias{addName,curv2Filter,logical-method}
\title{These methods are copied from flowViz to eliminate its dependency on curv1Filter and curv2Filter}
\usage{
\S4method{addName}{curv1Filter,character}(x, name, data, ...)

\S4method{addName}{curv1Filter,logical}(x, name, data, ...)

\S4method{addName}{curv2Filter,character}(x, name, data, ...)

\S4method{addName}{curv2Filter,logical}(x, name, data, ...)
}
\arguments{
\item{x}{curv1Filter, curv2Filter}

\item{name}{character or logical. Names can be generated by the filter or by the user.}

\item{data}{flowFrame}

\item{...}{other arguments}
}
\value{
The methods are called for their side effects. No value is returned.
}
\description{
These methods are copied from flowViz to eliminate its dependency on curv1Filter and curv2Filter
}
