% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregate_forecast_metero.R
\name{aggregate_forecast_metero}
\alias{aggregate_forecast_metero}
\title{Aggregate forecast meterological data from ICON-D2 and ICON-DE at forecast
time step to daily}
\usage{
aggregate_forecast_metero(forecast_data)
}
\arguments{
\item{forecast_data}{data frame of raw forecast meterological data with
forecasting time steps for different variables}
}
\value{
forecast_data aggregated to daily time steps
}
\description{
Aggregate forecast meterological data from ICON-D2 and ICON-DE at forecast
time step to daily
}
\examples{

\dontrun{
# Save downloaded data in this folder
data_dir <- "C:/Users/nguyenta/Documents/test"

Download for these variables
variables <- c("t_2m", "relhum_2m", "tot_prec")

# Now download the data (get_forecast_metero function in the R folder)
forecast_metadata  <- get_forecast_metero(data_dir, variables)

# Get basin shape file
basins <- sf::read_sf(file.path(.libPaths(),"FlowStats", "FlowStats","data",
                               "de_basins.shp"))

forecast_data <- extract_de_forecast(data_dir, forecast_metadata, basins)

forecast_data_agg <- aggregate_forecast_metero(forecast_data)

}
}
