% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flowstats.R
\name{calculate_flowstats}
\alias{calculate_flowstats}
\title{Calculate streamflow statistic}
\usage{
calculate_flowstats(
  Q_input,
  select_period,
  n_day_mean,
  stat_function,
  gui,
  flood_percentile = FALSE
)
}
\arguments{
\item{Q_input}{dataframe object, simulated or observed streamflow data, must
contains the following columns "gauge_id", "date", "q_mm_day"}

\item{select_period}{date vector with format "%Y-%m-%d", starting and ending
date of of the period for calculating streamflow statistic}

\item{n_day_mean}{smoothing the streamflow data using moving window, so the
streamflow of day t is the average streamflow from day t-n-1 to t}

\item{stat_function}{character, can be one of the following options
'Q_min (selected period)' for minimum, 'Q_mean (selected period)' to calculate
mean, 'Q_max (selected period)' to calculate maximum, or 'Q_last_day' to just
get Q latest day in the Q_input dataset}

\item{gui}{logical variable, by default is TRUE to if this function is called
inside the shiny serve to display the processing time. Otherwise when users
do not use the GUI, just set to FALSE}

\item{flood_percentile}{logical variable, FALSE if this is for flood classification,
otherwise TRUE}
}
\value{
a dataframe object, containing streamflow percentiles of the gaugeid
}
\description{
Calculate streamflow statistic
}
\examples{

\dontrun{
# Read data from FlowStats package
Q_input <- readRDS(file.path(.libPaths(),"FlowStats", "FlowStats",
                            "data", "lstm_data", "de_sim_discharge.rds"))


# Period for calculating streamflow statistic
select_period <- c(as.Date("2025-05-01"), as.Date("2025-05-31"))
n_day_mean <- 10
stat_function <- "Mean"
gui <- FALSE
flood_percentile <- FALSE

# Get flow percentiles
percentile <- calculate_flowstats(Q_input, select_period,
                                  n_day_mean, stat_function, gui)

}

}
