% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ellipse.R
\name{ellipse}
\alias{ellipse}
\title{convert ellipse from cov/mu to points
used to plot priors}
\usage{
ellipse(cov, centre, level = 0.95)
}
\description{
convert ellipse from cov/mu to points
used to plot priors
}
