% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_forecast.R
\name{extract_de_forecast}
\alias{extract_de_forecast}
\title{Extracting meteorological data downloaded from ICON-D2 and ICON-EU to basin level}
\usage{
extract_de_forecast(data_dir, forecast_metadata, basins)
}
\arguments{
\item{data_dir}{directory of downloaded data from ICON-D2 and ICON-EU files}

\item{forecast_metadata}{dataframe object of three columns, variable name,
predicted time (UTC), and downloaded file names}

\item{basins}{basin shape object}
}
\value{
a data frame of forecasted meterological data at the forecast time
step and for different basins
}
\description{
Extracting meteorological data downloaded from ICON-D2 and ICON-EU to basin level
}
\examples{
\dontrun{

data_dir <- tempdir()
basins <- sf::read_sf(file.path(.libPaths(),"FlowStats", "FlowStats",
                                "data", "de_basins.shp"))
forecast_metadata <- get_forecast_metero(data_dir , variables)
extract_de_forecast(data_dir, forecast_metadata, basins)

}
}
