% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mkPriors.R
\name{flowClust2Prior}
\alias{flowClust2Prior}
\title{Generate a prior specification based on a flowClust model This function
generates a prior specification based on a flowClust fit object It can be
passed to a second round of flowClust() with usePrior="yes" The prior could
be estimated from a single sample, for example, and then used to speed up
the convergence for other samples.}
\usage{
flowClust2Prior(x, kappa, Nt = NULL, addCluster = NULL)
}
\arguments{
\item{x}{a flowClust fit object}

\item{kappa}{is the fraction of equivalent observations by which to weight
this prior relative to the flowClust model.}

\item{Nt}{the number of total equivalent observation}

\item{addCluster}{not currently supported}
}
\description{
Generate a prior specification based on a flowClust model This function
generates a prior specification based on a flowClust fit object It can be
passed to a second round of flowClust() with usePrior="yes" The prior could
be estimated from a single sample, for example, and then used to speed up
the convergence for other samples.
}
