% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_forecast_metero.R
\name{get_forecast_metero}
\alias{get_forecast_metero}
\title{Download ICON-D2 (1-48 hours) and ICON-EU (49-120 hours) meteorological data}
\usage{
get_forecast_metero(data_dir, variables)
}
\arguments{
\item{data_dir}{directory to save downloaded data from ICON-D2 and ICON-EU
files}

\item{variables}{variable names, could be one or multiple variables (vector),
for example c("t_2m", "relhum_2m", "tot_prec"). A list of all variables names
can be found here https://opendata.dwd.de/weather/nwp/icon-d2/grib/00/}
}
\value{
a dataframe object of three columns, variable name, predicted time (UTC),
and downloaded file names
}
\description{
Download ICON-D2 (1-48 hours) and ICON-EU (49-120 hours) meteorological data
}
\examples{
\dontrun{

data_dir <- tempdir()
variables <- c("t_2m", "relhum_2m", "tot_prec")
forecast_files <- get_forecast_metero(data_dir , variables)

}
}
