% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_historical_metero.R
\name{get_historical_metero}
\alias{get_historical_metero}
\title{Get  most recent daily gridded HYDRAS hydrological data (including
precipitation, tempearture min and max, humidity) from the Deutsche
Wetterdienst}
\usage{
get_historical_metero(years = NA, data_dir = NA)
}
\arguments{
\item{years}{integer vector, list of all years that users want to download.
Each file of the HYDRAS data contains 1 calendar year. Since the base data
already have simulated streamflow from 1980 to 2025. Therefore, users only
need to download data from 2025}

\item{data_dir}{path to save downloaded data, if not given, the temporary path
will be used}
}
\value{
a list object, containing information about the downloaded files,
years, and path to the downloaded data
}
\description{
Get  most recent daily gridded HYDRAS hydrological data (including
precipitation, tempearture min and max, humidity) from the Deutsche
Wetterdienst
}
\examples{

\dontrun{
meterological_data <- get_historical_metero(years = NA, data_dir = NA)
}

}
