\name{ITN}
\alias{ITN}
\docType{data}
\title{ Sample flow cytometry data }
\description{
  A \code{\link[flowCore:flowSet-class]{flowSet}} cotaining data from 15
  patients.
}
\usage{data(ITN)}
\format{
  A \code{flowSet} containing 15
  \code{\link[flowCore:flowFrame-class]{flowFrames}}. There are 3
  patient groups with 5 samples each. 
}
\source{
 Immune Tolerance Network
}
\keyword{datasets}
