% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_flowstats.R
\name{plot_flowstats}
\alias{plot_flowstats}
\title{Plot streamflow statistic for one gauge}
\usage{
plot_flowstats(
  Q_input,
  select_period,
  n_day_mean,
  stat_function,
  percentile_class
)
}
\arguments{
\item{Q_input}{dataframe object, simulated or observed streamflow data, must
contains the following columns "gauge_id", "date", "q_mm_day". IMPORTANT, this
data frame should have only one gauge id}

\item{select_period}{date vector with format "%Y-%m-%d", starting and ending date of
of the period for calculating streamflow statistic}

\item{n_day_mean}{integer, number of days for calculating the moving average}

\item{stat_function}{character, could be "Min", "Max", or "Median"}

\item{percentile_class}{percentile of streamflow should be calculated to
higlight all ranges of flows "All" or only for "Drought" or only for "Flood"}
}
\value{
ggplot object
}
\description{
Plot streamflow statistic for one gauge
}
\examples{
\dontrun{
select_period <- c(as.Date("2025-05-01"), as.Date("2025-05-31"))
stat_function <- "Mean"
n_day_mean <- 1
percentile_class <- "All"
}
}
