% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_gauge_stats.R
\name{plot_gauge_stats}
\alias{plot_gauge_stats}
\title{Plot streamflow statistic for one gauge}
\usage{
plot_gauge_stats(Q_input, period, gaugeid, fun)
}
\arguments{
\item{Q_input}{dataframe object, simulated or observed streamflow data, must
contains the following columns "gauge_id", "date", "Q_cms"}

\item{period}{date vector with format "%Y-%m-%d", starting and ending date of
of the period for calculating streamflow statistic}

\item{gaugeid}{character vector, list of gauge_id which users want to
calculate streamflow statistics}

\item{fun}{character, can be one of the following options
'Q_min (selected period)' for minimum, 'Q_mean (selected period)' to calculate
mean, 'Q_max (selected period)' to calculate maximum, or 'Q_last_day' to just
get Q latest day in the Q_input dataset}
}
\value{
ggplot object
}
\description{
Plot streamflow statistic for one gauge
}
\examples{
\dontrun{
# Read data from FlowStats package
Q_input <- readRDS(file.path(.libPaths(),"FlowStats", "FlowStats",
                            "data", "lstm_data", "de_sim_discharge.rds"))

# Period for calculating streamflow statistic
period <- c(as.Date("2025-05-01"), as.Date("2025-05-31"))

# Gauge id
gaugeid <- Q_input$gauge_id[1]

# How to calculate Q
fun <- "Q_mean (selected period)"

# Get flow percentiles
plt_gauge_stats <- plot_gauge_stats(Q_input, period, gaugeid, fun)
}
}
