% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_timeseries.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot streamflow statistic for one gauge}
\usage{
plot_timeseries(Q_input, gaugeid, plot_type, log_y)
}
\arguments{
\item{Q_input}{dataframe object, simulated or observed streamflow data, must
contains the following columns "gauge_id", "date", "q_mm_day"}

\item{gaugeid}{character, selected gauge_id}

\item{plot_type}{character, can be one of the following options 'Daily',
'Daily (by year)' for minimum, 'Daily cumsum (by year)'}

\item{log_y}{1 for log_y axis and 0 for normal axis}
}
\value{
ggplot object
}
\description{
Plot streamflow statistic for one gauge
}
\examples{
\dontrun{
# Read data from FlowStats package
Q_input <- readRDS(file.path(.libPaths(),"FlowStats", "FlowStats",
                            "data", "lstm_data", "de_sim_discharge.rds"))

# Period for calculating streamflow statistic
period <- c(as.Date("2025-05-01"), as.Date("2025-05-31"))

# Gauge id
gaugeid <- Q_input$gauge_id[1]

# How to calculate Q
plot_type <- "Daily (by year)"

# Get flow percentiles
plt_timeseries <- plot_timeseries(Q_input, gaugeid, plot_type, log_y = FALSE)
}
}
