% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_flowstats.R
\name{show_flowstats}
\alias{show_flowstats}
\title{Display the Graphical User Interface (GUI) of FlowStats}
\usage{
show_flowstats()
}
\value{
GUI of FlowStats
}
\description{
Calling this functions will show the GUI of FlowStats
}
\examples{

\dontrun{
show_flowstats()
}

}
