% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/showFlowStats.R
\name{showFlowStats}
\alias{showFlowStats}
\title{Display the Graphical User Interface (GUI) of FlowStats}
\usage{
showFlowStats()
}
\value{
GUI of FlowStats
}
\description{
Calling this functions will show the GUI of FlowStats
}
\examples{

\donttest{
showFlowStats()
}

}
