% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_flowstats.R
\name{calculate_flowstats}
\alias{calculate_flowstats}
\title{Calculate streamflow statistic}
\usage{
calculate_flowstats(Q_input, period, gaugeid, fun, gui = TRUE)
}
\arguments{
\item{Q_input}{dataframe object, simulated or observed streamflow data, must
contains the following columns "gauge_id", "date", "Q_cms"}

\item{period}{date vector with format "%Y-%m-%d", starting and ending date of
of the period for calculating streamflow statistic}

\item{gaugeid}{character vector, list of gauge_id which users want to
calculate streamflow statistics}

\item{fun}{character, can be one of the following options
'Q_min (selected period)' for minimum, 'Q_mean (selected period)' to calculate
mean, 'Q_max (selected period)' to calculate maximum, or 'Q_last_day' to just
get Q latest day in the Q_input dataset}

\item{gui}{logical variable, by default is TRUE to if this function is called
inside the shiny serve to display the processing time. Otherwise when users
do not use the GUI, just set to FALSE}
}
\value{
a dataframe object, containing streamflow percentiles of the gaugeid
}
\description{
Calculate streamflow statistic
}
\examples{
\dontrun{
# Read data from FlowStats package
Q_input <- readRDS(file.path(.libPaths(),"FlowStats", "FlowStats",
                            "data", "lstm_data", "de_sim_discharge.rds"))


# Period for calculating streamflow statistic
period <- c(as.Date("2025-05-01"), as.Date("2025-05-31"))

# Gauge id
gaugeid <- Q_input$gauge_id[1]

# How to calculate Q
plot_type <- "Daily (by year)"

# Get flow percentiles
flow_stats <- calculate_flowstats(Q_input, period, gaugeid,
                                  fun = "Q_min (selected period)",
                                  gui = FALSE)
}
}
