% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fda_functions.R
\name{fdPar}
\alias{fdPar}
\title{the old version copied from fda 2.4.0 because the new API changes the output 
(specifically res$fd$coefs) and thus breaks the landmarkreg call}
\usage{
fdPar(fdobj = NULL, Lfdobj = NULL, lambda = 0, estimate = TRUE,
  penmat = NULL)
}
\description{
the old version copied from fda 2.4.0 because the new API changes the output 
(specifically res$fd$coefs) and thus breaks the landmarkreg call
}
