\name{flowStats-package}

\alias{flowStats-package}

\alias{flowStats}

\docType{package}

\title{
  
Statistical methods for flow cytometry data analysis

}

\description{
  
Functions, methods and classes implementing algorithmns for statistical
analysis of flow cytometry data. This involves mostly data normalization
and automated gating.

}

\details{
  
  \tabular{ll}{
    Package: \tab flowStats\cr
    Type: \tab Package\cr
    Version: \tab 1.0\cr
    License: \tab Artistic-2.0\cr
    Lazyload: \tab yes\cr
  }

}

\author{
  
Florian Hahne

Maintainers: Mike Jiang <mike@ozette.ai>, Jake Wagner <jpwagner@fhcrc.org>

}


\keyword{ package }

