% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grid_to_basins.R
\name{grid_to_basin}
\alias{grid_to_basin}
\title{Extract gridded HYDRAS hydrological data and aggregate at the basins levels}
\usage{
grid_to_basin(years, data_dir, basins)
}
\arguments{
\item{years}{integer vector, list of all years of the downloaded HYDRAS data}

\item{data_dir}{path to save extracted data}

\item{basins}{shape file object of basins}
}
\value{
a list object, containing extract data for each basin
}
\description{
Extract gridded HYDRAS hydrological data and aggregate at the basins levels
}
\examples{

\donttest{
basins <- read_sf("your_path/de_basins.shp")
basin_data <- grid_to_basin(years = c(2024:2025), data_dir = tempdir(),
                           basins = basins)
}

}
