% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_data.R
\name{update_data}
\alias{update_data}
\title{Calculate streamflow statistic}
\usage{
update_data(env_path, basins, forecast)
}
\arguments{
\item{env_path}{path to the python virtual envionment to run the LSTM model.
The python virtual environment can be created using annaconda with the
environment file (environment.yml, link to this file is in the example)
in the package folder. The LSTM model was constructed using the HydroEcoLSTM
package (https://github.com/tamnva/hydroecolstm).}

\item{basins}{basin shape file from the CAMEL-DE dataset (Loritz et al., 2024)}

\item{forecast}{logical variable, TRUE meaning that forecast meteorological
data will be downloaded, FALSE if use only want to update historical
meterological data}
}
\value{
the base data will be updated
}
\description{
Calculate streamflow statistic
}
\examples{

\dontrun{
# The path of Python environment for running the LSTM model
# Here is the path to the environment.yml file using annaconda
# More information can be found at https://github.com/tamnva/hydroecolstm
file.path(.libPaths(), "FlowStats", "FlowStats","data",
                            "lstm_data", "environment.yml")

env_path <- "C:/Users/nguyenta/AppData/Local/anaconda3/envs/hydro"

# Read the basin shape file
basins <- sf::read_sf(file.path(.libPaths(),"FlowStats", "FlowStats",
                               "data", "de_basins.shp"))

# Update the data
update_data(env_path, basins, forecast = TRUE)

# Here is the updated data file
Q_data <- readRDS(file.path(.libPaths(), "FlowStats", "FlowStats","data",
                            "lstm_data", "de_sim_discharge.rds"))

}

}
