% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_base_data.R
\name{update_data}
\alias{update_data}
\title{Update existing meterological data with near-real-time (1-day delay)
meterological data from Deutsche Wetterdienst}
\usage{
update_data(old_data_file, basins, years = NA, data_dir = NA)
}
\arguments{
\item{old_data_file}{character: path to old data file.}

\item{basins}{shape file object of basins}

\item{years}{integer vector, list of all years that users want to download.
Each file of the HYDRAS data contains 1 calendar year. Since the base data
already have simulated streamflow from 1980 to 2025. Therefore, users only
need to download data from 2025}

\item{data_dir}{path to save downloaded data, if not given, the temporary path
will be used}
}
\value{
new data will be added to the old data. IMPORTANT: The old_data_file
will be replace
}
\description{
Update existing meterological data with near-real-time (1-day delay)
meterological data from Deutsche Wetterdienst
}
\examples{

\donttest{
old_data_file <- "your_path/time_series.csv"
basins <- read_sf("your_path/de_basins.shp")
update_data(old_data_file, basins, years = NA, data_dir = NA)


}
}
